requirements.txt
# ── LiveKit ──────────────────────────────────────────────────────────
livekit==1.1.2
livekit-agents==1.1.7
livekit-plugins-silero==1.1.7

# ── Speech-to-Text ───────────────────────────────────────────────────
deepgram-sdk==3.5.0

# ── Text-to-Speech ───────────────────────────────────────────────────
gTTS==2.5.4
pydub==0.25.1

# ── Audio / Numeric ──────────────────────────────────────────────────
numpy==1.26.4

# ── WebSockets (Deepgram transport) ──────────────────────────────────
websockets==12.0

# ── Config ───────────────────────────────────────────────────────────
python-dotenv==1.0.1
